refactor(rcbc): move data *.ts files to a dedicated folder
