+++ scripts/typescript/src/rcbc/data/materials.ts
