+++ scripts/typescript/src/rcbc/data/shared.ts
